pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-gateway"
version = "0.1.0"
description = "REST API gateway for Model Context Protocol servers"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "axios contributors" }
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
    "pydantic>=2.0.0",
    "httpx>=0.26.0",
    "mcp>=1.0.0",
    "sse-starlette>=2.0.0",
    # OAuth2 authentication
    "authlib>=1.3.0",
    "itsdangerous>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
]

[project.scripts]
mcp-gateway = "mcp_gateway.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_gateway"]
